requirements.txt
torchtext==0.6.0
pytorch==1.2.0
numpy==1.18.5
transformers==2.11.0
nltk==3.5
spacy==2.3.0
